create row as a new component and exclude it from app
